Gives a circular aparence to the SVG icon in the menu.
The Aeskulap vector icon is circular, but in the SVG file it is in a
rectangular field, that is coerced into a square and gives the icon an
elliptical shape in the desktop menu. This patch was made with Inkscape
by selecting the icon, and adjusting the document size to the icon. This
restores the circular appearance of the icon in the desktop menu.

Bug: https://savannah.nongnu.org/bugs/index.php?29103